% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ist_categorical}
\alias{ist_categorical}
\title{Paleta categorial (12 cores)}
\usage{
ist_categorical()
}
\value{
Vector de 12 caracteres hexadecimais.
}
\description{
Paleta categorial (12 cores)
}
\examples{
ist_categorical()
}
