% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ist_colours}
\alias{ist_colours}
\title{Cores institucionais principais}
\usage{
ist_colours()
}
\value{
Vector nomeado de 3 hexadecimais.
}
\description{
Cores institucionais principais
}
\examples{
ist_colours()
ist_colours()[["blue"]]
}
