% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{ist_font_path}
\alias{ist_font_path}
\title{Caminho para a fonte Source Sans 3 embutida no pacote}
\usage{
ist_font_path()
}
\value{
String com o caminho para o ficheiro .ttf, ou "" se não encontrado.
}
\description{
Caminho para a fonte Source Sans 3 embutida no pacote
}
\keyword{internal}
