% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ist_sequential}
\alias{ist_sequential}
\title{Escala sequencial (12 tons do azul principal)}
\usage{
ist_sequential()
}
\value{
Vector de 12 caracteres hexadecimais (claro -> escuro).
}
\description{
Escala sequencial (12 tons do azul principal)
}
\examples{
ist_sequential()
}
