% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_ist}
\alias{scale_colour_ist}
\alias{scale_color_ist}
\alias{scale_fill_ist}
\title{Escala de cor categorial institucional (discreta)}
\usage{
scale_colour_ist(...)

scale_color_ist(...)

scale_fill_ist(...)
}
\arguments{
\item{...}{Argumentos passados para \code{\link[ggplot2]{discrete_scale}}.}
}
\description{
Escala de cor categorial institucional (discreta)
}
