% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_ist_seq}
\alias{scale_colour_ist_seq}
\alias{scale_color_ist_seq}
\alias{scale_fill_ist_seq}
\title{Escala de cor sequencial institucional (contínua)}
\usage{
scale_colour_ist_seq(...)

scale_color_ist_seq(...)

scale_fill_ist_seq(...)
}
\arguments{
\item{...}{Argumentos passados para \code{\link[ggplot2]{continuous_scale}}.}
}
\description{
Escala de cor sequencial institucional (contínua)
}
